chore: moved apple specific code out of shared files.c
